+++ projects/sds/src/app.routes.ts
@@ -1,8 +1,10 @@
 import {Routes, UrlMatchResult, UrlSegment} from '@angular/router';
 import {URL_CONST} from './app/shared/url/url.const';
+import {InvalidComponent} from './app/invalid/invalid.component';
 
 export const APP_ROUTES: Routes = [
 	{path: '', redirectTo: 'introductions/welcome', pathMatch: 'full'},
+	{path: 'invalid', component: InvalidComponent},
 	{
 		loadChildren: () => import('./app/component-page/component-pages.routes'),
 		matcher: (url: UrlSegment[]): UrlMatchResult =>
@@ -30,6 +32,6 @@ export const APP_ROUTES: Routes = [
 	},
 	{
 		path: '**',
-		redirectTo: 'introductions/welcome'
+		redirectTo: 'invalid'
 	}
 ];