+++ projects/sds/src/app/component-page/component-pages.routes.ts
@@ -2,5 +2,5 @@ import {ComponentPageComponent} from './component-page.component';
 
 export default [
 	{path: 'newsletter', component: ComponentPageComponent},
-	{path: '**', redirectTo: 'introductions/welcome'}
+	{path: '**', redirectTo: '../invalid'}
 ];