+++ projects/sds/src/app/text-page/text-pages.routes.ts
@@ -2,5 +2,5 @@ import {TextPageComponent} from './text-page.component';
 
 export default [
 	{path: '', component: TextPageComponent},
-	{path: '**', redirectTo: 'introductions/welcome'}
+	{path: '**', redirectTo: '../invalid'}
 ];